feat: static view filter demo refactor[WTEL-6934](https://webitel.atlassian.net/browse/WTEL-6934)
