+++ src/modules/configuration/modules/lookups/modules/contact-groups/modules/add-contacts-in-group/configs/filtersOptions.ts
@@ -1,0 +1,14 @@
+import {
+  createFilterConfig,
+  FilterConfigDefinition,
+  FilterOption,
+} from '@webitel/ui-datalist/filters';
+
+export const filtersOptions: FilterConfigDefinition[] = [
+  createFilterConfig({
+    name: FilterOption.ContactGroup,
+    hideUnassigned: true,
+  }),
+  FilterOption.HasUser,
+  FilterOption.ContactLabel,
+];
