+++ src/modules/configuration/modules/lookups/modules/contact-groups/modules/add-contacts-in-group/namespace.ts
@@ -1,0 +1,3 @@
+import { ContactGroupsNamespace } from '../../namespace';
+
+export const AddContactsInGroupNamespace = `${ContactGroupsNamespace}/add-contacts-in-group`;
